+++ server/impl/src/main/java/com/walmartlabs/concord/server/agent/AgentResource.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.agent;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,11 +20,13 @@ package com.walmartlabs.concord.server.agent;
  * =====
  */
 
+import com.walmartlabs.concord.common.ConfigurationUtils;
 import com.walmartlabs.concord.server.sdk.metrics.WithTimer;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.Authorization;
 import org.sonatype.siesta.Resource;
+import org.sonatype.siesta.ValidationErrorsException;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -32,8 +34,14 @@ import javax.inject.Singleton;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Named
 @Singleton
@@ -55,5 +63,30 @@ public class AgentResource implements Resource {
     @WithTimer
     public Collection<AgentWorkerEntry> listWorkers() {
         return agentManager.getAvailableAgents();
+    }
+
+    @GET
+    @ApiOperation(value = "Counts the currently connected workers based on the specified capabilities property")
+    @Path("/all/workersCount")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Map<Object, Long> aggregate(@QueryParam("capabilities") String capabilities) {
+        if (capabilities == null || capabilities.isEmpty()) {
+            throw new ValidationErrorsException("'capabilities' filter is required");
+        }
+
+        Collection<AgentWorkerEntry> data = agentManager.getAvailableAgents();
+        if (data == null || data.isEmpty()) {
+            return Collections.emptyMap();
+        }
+
+        String[] path = capabilities.split("\\.");
+        if (path.length < 1 || Stream.of(path).anyMatch(p -> p.trim().isEmpty())) {
+            throw new ValidationErrorsException("Invalid 'capabilities' value. Expected a path to a property, got: " + capabilities);
+        }
+
+        return data.stream()
+                .map(e -> ConfigurationUtils.get(e.capabilities(), path))
+                .map(e -> e != null ? e : "n/a")
+                .collect(Collectors.groupingBy(Function.identity(), Collectors.counting()));
     }
 }